MorseGame/Visualization and Additional fixes
In Morse Game, Added visualization, which is basically drawing to the
screen. It can easily be expanded for other shapes HOWEVER these are
STATIC elements and NOT DYNAMIC (for now).

Also fixes up a few things, about converting to morse_code, time between
Morse problems, and good transition between them, also correct checking
for answers.

This needs further work with Signals when putting it all together, but
should be simple enough.